requirements.txt
hermes_python>=0.1
RPi.Gpio